pyproject.toml
[project]
name = "murphy"
description = "AI-driven website evaluation powered by browser-use"
authors = [{ name = "MIH AI B.V." }]
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
# Murphy's own imports are lightweight (browser-use-sdk, uuid7, click, rich,
# pydantic, fastapi, etc.). The remaining packages below are required by the
# vendored browser_use/ engine and its LLM provider integrations.
dependencies = [
    "aiohttp>=3.13.3",
    "anyio>=4.9.0",
    "bubus>=1.5.6",
    "click>=8.1.8",
    "InquirerPy>=0.3.4",
    "rich>=14.0.0",
    "google-api-core>=2.25.0",
    "httpx>=0.28.1",
    "portalocker>=2.7.0,<3.0.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.10.1",
    "pyobjc>=11.0; platform_system == 'darwin'",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "screeninfo>=0.8.1; platform_system != 'darwin'",
    "typing-extensions>=4.12.2",
    "uuid7>=0.1.0",
    "authlib>=1.6.0",
    "google-genai>=1.50.0,<2.0.0",
    "openai>=2.7.2,<3.0.0",
    "anthropic>=0.72.1,<1.0.0",
    "groq>=0.30.0",
    "ollama>=0.5.1",
    "google-api-python-client>=2.174.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "pypdf>=5.7.0",
    "reportlab>=4.0.0",
    "cdp-use>=1.4.4",
    "pyotp>=2.9.0",
    "pillow>=11.2.1",
    "cloudpickle>=3.1.1",
    "markdownify>=1.2.0",
    "python-docx>=1.2.0",
    "browser-use-sdk>=2.0.12",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
video = ["imageio[ffmpeg]>=2.37.0", "numpy>=2.3.2"]
eval = [
    "lmnr[all]==0.7.17",
    "anyio>=4.9.0",
    "psutil>=7.0.0",
    "datamodel-code-generator>=0.26.0",
]


[project.urls]
Repository = "https://github.com/ProsusAI/Murphy"

[project.scripts]
murphy = "murphy.api.cli:main"  # AI-driven website evaluation (primary)
murphy-api = "murphy.api.rest:main"  # Murphy REST API server

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.codespell]
ignore-words-list = "bu,wit,dont,cant,wont,re-use,re-used,re-using,re-usable,thats,doesnt,doubleclick,finaly,finalY"
skip = "*.json"

[tool.ruff]
line-length = 130
fix = true

[tool.ruff.lint]
select = ["ASYNC", "E", "F", "FAST", "I", "PLE"]
ignore = [
    "ASYNC109",
    "ASYNC240",  # sync Path/os.path in async — acceptable for quick local file ops
    "ASYNC250",  # blocking input() in async — intentional for interactive CLI prompts
    "E101",
    "E402",
    "E501",
    "W291",
]
unfixable = ["E101", "E402", "E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
line-ending = "lf"
docstring-code-format = true
docstring-code-line-length = 140
skip-magic-trailing-comma = false

[tool.pyright]
typeCheckingMode = "basic"
include = ["murphy", "browser_use", "tests"]
exclude = [
    ".venv/",
    ".venv*/",
    ".git/",
    "__pycache__/",
    "**/site-packages/",
    "./test_*.py",
    "./debug_*.py",
    "private_example/",
    "debug/*",
    "tests/scripts/*",
    "tests/old/*",
    "browser_use/dom/playground/*",
    "browser_use/llm/oci_raw/*",
    "browser_use/llm/tests/test_chat_models.py",
    "browser_use/llm/tests/test_single_step.py",
    "product_extraction.py",
    "discover/",
    "list/",
]
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false


[tool.hatch.build]
include = [
    "murphy/**/*.py",
    "browser_use/**/*.py",
    "!browser_use/**/tests/*.py",
    "!browser_use/**/tests.py",
	"browser_use/agent/system_prompts/*.md",
    "browser_use/py.typed",
    "browser_use/dom/**/*.js",
    "!tests/**/*.py",
    "!debug/*",
]

[tool.pytest.ini_options]
timeout = 300
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "slow: marks tests as slow (deselect with `-m 'not slow'`)",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: mark tests as async tests",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-svx --strict-markers --tb=short --dist=loadscope"
log_cli = true
log_cli_format = "%(levelname)-8s [%(name)s] %(message)s"
filterwarnings = [
    "ignore::pytest.PytestDeprecationWarning",
    "ignore::DeprecationWarning",
]
log_level = "DEBUG"


[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
# required-environments = [
#     "sys_platform == 'darwin' and platform_machine == 'arm64'",
#     "sys_platform == 'darwin' and platform_machine == 'x86_64'",
#     "sys_platform == 'linux' and platform_machine == 'x86_64'",
#     "sys_platform == 'linux' and platform_machine == 'aarch64'",
#     # "sys_platform == 'linux' and platform_machine == 'arm64'",  # no pytorch wheels available yet
#     "sys_platform == 'win32' and platform_machine == 'x86_64'",
#     # "sys_platform == 'win32' and platform_machine == 'arm64'",  # no pytorch wheels available yet
# ]
dev-dependencies = [
    "ruff>=0.11.2",
    "tokencost>=0.1.16",
    "build>=1.2.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-httpserver>=1.0.8",
    "fastapi>=0.115.8",
    "inngest>=0.4.19",
    "uvicorn>=0.34.0",
    "ipdb>=0.13.13",
    "pre-commit>=4.2.0",
    "codespell>=2.4.1",
    "pyright>=1.1.403",
    "ty>=0.0.1a1",
    "pytest-xdist>=3.7.0",
    "lmnr[all]==0.7.17",
    # "pytest-playwright-asyncio>=0.7.0",  # not actually needed I think
    "pytest-timeout>=2.4.0",
    "pytest-cov>=7.0.0",
]
